[#56] [Fix] 프로젝트 게시판 수정

+++ src/main/java/com/example/dopamines/domain/board/project/repository/ProjectPostRepository.java
@@ -10,5 +10,5 @@ import java.util.Optional;
 @Repository
 public interface ProjectPostRepository extends JpaRepository<ProjectPost, Long> {
 
-    Optional<List<ProjectPost>> findByCourseNum(Long courseNum);
+    Optional<List<ProjectPost>> findByCourseNum(Integer courseNum);
 }
